Fix flow control for SPDY v3
SPDY v3.1 introduces the concept of session-level flow control. Prior to that,
SPDY v3 only implemented stream-level flow control. This commit fixes the Stream
class, which was incorrectly performing flow-control at both the stream and the
session level, to only deal with stream flow-control for SPDY v3, with optional
support for SPDY v3.1, if the settings mandate it.