Fix auto-upgrade from SPDY/3 to SPDY/3.1
The current code auto-upgrades all SPDY/3 clients to SPDY/3.1 when
operating in **plain** mode with **autoSpdy31** set to true. This seems
rather restrictive, implying that with autoSpdy31 set to true, the
server does not support SPDY/3-only clients.

This commit fixes this by only upgrading SPDY/3 clients to SPDY/3.1 when
the server receives a WINDOW_UPDATE for stream id 0. This ensures
that SPDY/3 clients can proceed uninhibited when operating in **plain**
mode, and ensuring that the autoSpdy31 option is only used for
SPDY/3.1 clients in **plain** mode, allowing for both SPDY/3 and
SPDY/3.1 clients to co-exist on the same server in **plain** mode.